requirements.txt
# Core geospatial
rasterio>=1.3.0
GDAL>=3.4.0
geopandas>=0.13.0
shapely>=2.0.0
pyproj>=3.5.0
xarray>=2023.1.0
rioxarray>=0.14.0

# Scalable processing
dask[complete]>=2023.5.0
distributed>=2023.5.0

# ML / Deep Learning
torch>=2.0.0
torchvision>=0.15.0
segmentation-models-pytorch>=0.3.3
albumentations>=1.3.0

# Data & numerics
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.2.0
Pillow>=9.5.0
opencv-python>=4.7.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
folium>=0.14.0

# Config & utilities
PyYAML>=6.0
tqdm>=4.65.0
click>=8.1.0
loguru>=0.7.0
omegaconf>=2.3.0

# Testing
pytest>=7.3.0
pytest-cov>=4.0.0

# Notebooks
jupyter>=1.0.0
ipywidgets>=8.0.0
